requirements.txt
cx-Oracle==8.3.0
pandas
sqlalchemy
python-dotenv